update from https://github.com/CyanogenMod/android_system_core/commits/0fcbde444948c6851d08a2c4de0a44d555f112f5/mkbootimg

+++ src/unpackbootimg.c
@@ -16,7 +16,7 @@ int read_padding(FILE* f, unsigned itemsize, int pagesize)
 {
     byte* buf = (byte*)malloc(sizeof(byte) * pagesize);
     unsigned pagemask = pagesize - 1;
-    ssize_t count;
+    unsigned count;
 
     if((itemsize & pagemask) == 0) {
         free(buf);
